requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
colorama==0.4.6
coverage==7.3.2
cryptography==41.0.5
Django==4.2.6
djangorestframework==3.14.0
idna==3.4
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pycparser==2.21
PyMySQL==1.1.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-django==4.6.0
pytest-mock==3.12.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
